+++ database/migrations/2023_10_11_045132_create_doctor_infos_table.php
@@ -30,8 +30,8 @@ return new class extends Migration
             $table->string('commune_id')->nullable();
             $table->string('detail_address')->nullable();
 
-            $table->unsignedInteger('created_by')->nullable();
-            $table->unsignedInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
             $table->timestamps();
         });
     }
