+++ database/migrations/2023_10_11_080723_create_company_user_applieds_table.php
@@ -14,9 +14,9 @@ return new class extends Migration
         Schema::create('company_user_applieds', function (Blueprint $table) {
             $table->id();
 
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('company_id');
-            $table->unsignedInteger('post_recruitment_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('company_id');
+            $table->unsignedBigInteger('post_recruitment_id');
             $table->integer('status_applied')->default('1');
 
             $table->timestamps();
