+++ database/migrations/2023_10_11_082344_create_provinces_table.php
@@ -21,12 +21,12 @@ return new class extends Migration
             $table->string('full_name_en')->nullable();
             $table->string('code_name')->nullable();
 
-            $table->unsignedInteger('nation_id')->nullable();
-            $table->unsignedInteger('administrative_unit_id')->nullable();
-            $table->unsignedInteger('administrative_region_id')->nullable();
+            $table->unsignedBigInteger('nation_id')->nullable();
+            $table->unsignedBigInteger('administrative_unit_id')->nullable();
+            $table->unsignedBigInteger('administrative_region_id')->nullable();
 
-            $table->unsignedInteger('created_by')->nullable();
-            $table->unsignedInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
             $table->timestamps();
         });
     }
