+++ database/migrations/2023_10_11_082351_create_districts_table.php
@@ -22,10 +22,10 @@ return new class extends Migration
             $table->string('code_name')->nullable();
             $table->string('province_code')->nullable();
 
-            $table->unsignedInteger('administrative_unit_id')->nullable();
+            $table->unsignedBigInteger('administrative_unit_id')->nullable();
 
-            $table->unsignedInteger('created_by')->nullable();
-            $table->unsignedInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
             $table->timestamps();
         });
     }
